+++ meteor-app/imports/ui/service/tag-list.js
@@ -1,73 +1,64 @@
 import React from 'react'
 import PropTypes from 'prop-types'
-import './tag-list.css'
+import Tag from './tag'
+import { Label, Icon, Header } from 'semantic-ui-react'
 
-export default function TagList(props) {
-  const { removeTag, toggleTag, majorMinorTotal, totalPrice, toggleExpand, tags } = props
-
+export default TagList = ({
+  removeTag,
+  toggleTag,
+  changeTagName,
+  majorMinorTotal,
+  totalCost,
+  toggleExpand,
+  tags,
+  adjustPrice,
+}) => {
   return (
-    <div className="wrapper">
-      <div className="tag-wrapper">
-        {tags.map((tag, index) =>
-          tag.name !== 'Major Service' && tag.name !== 'Minor Service' ? (
-            <span className="tag" key={index}>
-              <span className="item-name" key={'a'}>
-                {tag.name}{' '}
-              </span>
-              <span className="item-name" key={'b'}>
-                &nbsp;${tag.price}
-              </span>
-              <span className="handle" key={'c'} onClick={() => removeTag(tag, index)}>
-                x
-              </span>
-            </span>
-          ) : null
+    <div id="pill-container" style={{ marginTop: '10px' }}>
+      {tags &&
+        tags.map((tag, index) =>
+          tag.items ? (
+            <>
+              <Label id="item-pill" key={index} style={{ margin: '5px' }} size="big" color="blue">
+                <Icon
+                  id="collapse-icon"
+                  name={tag.expanded ? 'angle double left' : 'angle double right'}
+                  style={{ cursor: 'pointer' }}
+                  onClick={() => toggleExpand(tag, index)}
+                />
+                {tag.name}
+                <Label.Detail>${majorMinorTotal(tag.items)}</Label.Detail>
+                <Icon id="delete-icon" name="delete" onClick={() => removeTag(tag, index)} />
+              </Label>
+              {tag.expanded &&
+                tag.items.map((item) => (
+                  <Label
+                    id="item-pill-small"
+                    key={item.name}
+                    style={{ margin: '5px' }}
+                    size="large"
+                    color={item.greyed ? null : 'teal'}
+                  >
+                    {item.name}
+                    <Label.Detail>${item.price}</Label.Detail>
+                    <Label.Detail>
+                      <Icon
+                        id="plus-icon"
+                        name={item.greyed ? 'plus' : 'minus'}
+                        style={{ cursor: 'pointer' }}
+                        onClick={() => toggleTag(item, tag)}
+                      />
+                    </Label.Detail>
+                  </Label>
+                ))}
+            </>
+          ) : (
+            <Tag index={index} tag={tag} adjustPrice={adjustPrice} newTagName={changeTagName} removeTag={removeTag} />
+          )
         )}
-      </div>
-
-      {tags.map((tag, index) =>
-        tag.name === 'Major Service' || tag.name === 'Minor Service' ? (
-          <div className="mjmn" key={`minor${index}`}>
-            <span className="tag">
-              <span className="expand" key={'a'} onClick={() => toggleExpand(tag, index)}>
-                {tag.expanded ? '<' : '>'}
-              </span>
-              {tag.name}
-              <span className="item-name" key={'b'}>
-                &nbsp;${majorMinorTotal(tag.items)}
-              </span>
-              <span className="handle" key={'c'} onClick={() => removeTag(tag, index)}>
-                x
-              </span>
-            </span>
-            {tag.expanded &&
-              tag.items.map((item, index) =>
-                tag.name === 'Major Service' ? (
-                  <span className="tag" key={`major${index}`} style={item.greyed ? { background: 'grey' } : {}}>
-                    {item.name}
-                    <span className="item-price" key={'a'}>
-                      &nbsp;${item.price}
-                    </span>
-                    <span className="handle" key={'b'} onClick={() => toggleTag(item, tag)}>
-                      {item.greyed ? '+' : '-'}
-                    </span>
-                  </span>
-                ) : (
-                  <span className="tag" key={index} style={item.greyed ? { background: 'grey' } : {}}>
-                    {item.name}
-                    <span className="item-price" key={'a'}>
-                      &nbsp;${item.price}
-                    </span>
-                    <span className="handle" key={'b'} onClick={() => toggleTag(item, tag)}>
-                      {item.greyed ? 'O' : 'X'}
-                    </span>
-                  </span>
-                )
-              )}
-          </div>
-        ) : null
-      )}
-      <div className="total-price">${totalPrice}</div>
+      <Header style={{ margin: '10px' }} id="total-price">
+        {'Total: $' + totalCost}
+      </Header>
     </div>
   )
 }
@@ -76,6 +67,6 @@ TagList.propTypes = {
   removeTag: PropTypes.func.isRequired,
   toggleTag: PropTypes.func.isRequired,
   majorMinorTotal: PropTypes.func.isRequired,
-  totalPrice: PropTypes.number.isRequired,
-  tags: PropTypes.array.isRequired
+  totalCost: PropTypes.number.isRequired,
+  tags: PropTypes.array.isRequired,
 }